add backgraun in signUpPage
