+++ src/pages/signUpPage/SignUpPage.css
@@ -1,3 +1,4 @@
 .backButton :hover {
     color: white; 
 }
+
